project/requirements.txt
asgiref==3.2.10
bcrypt==3.1.7
certifi==2020.12.5
cffi==1.14.1
chardet==4.0.0
ClickUz==0.3
decorator==4.4.2
Django==3.0.9
django-ckeditor==6.0.0
django-classy-tags==2.0.0
django-cms==3.8.0
django-cors-headers==3.7.0
django-decorator-include==3.0
django-filer==2.0.2
django-formtools==2.2
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==3.0.0
django-ratelimit==3.0.1
django-resized==0.3.11
django-restframework==0.0.1
django-sekizai==2.0.0
django-treebeard==4.3.1
djangocms-admin-style==2.0.1
djangocms-attributes-field==2.0.0
djangocms-bootstrap4==2.0.0
djangocms-icon==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-text-ckeditor==4.0.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.6.0
easy-thumbnails==2.7
html5lib==1.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
moviepy==1.0.3
mysqlclient==2.0.3
numpy==1.21.2
PaycomUz==2.2
Pillow==8.3.1
proglog==0.1.9
pycparser==2.20
PyJWT==1.7.1
pyTelegramBotAPI==3.7.5
python-dotenv==0.18.0
python-ffmpeg-video-streaming==0.1.14
pytz==2020.1
requests==2.25.1
six==1.15.0
sqlparse==0.3.1
tqdm==4.62.1
transliterate==1.10.2
Unidecode==1.1.1
urllib3==1.26.2
webencodings==0.5.1
